chore: cleaned up some code for harm🥺
