+++ Cargo.toml
@@ -4,6 +4,7 @@ edition = "2021"
 version = "1.0.2"
 rust-version = "1.65"
 repository = "https://github.com/d2foundry/oracle_engine/"
+# build = "build-script-src/main.rs"
 build = "build.rs"
 
 [build-dependencies]
@@ -12,6 +13,7 @@ serde = { version = "^1.0", features = ["derive"]}
 reqwest = { version = "^0.11", features = ["json", "blocking"] }
 serde_json = "^1.0"
 ron = "^0.8"
+thiserror = "1.0.48"
 
 [dev-dependencies]
 num-traits = "0.2"
@@ -20,13 +22,14 @@ num-traits = "0.2"
 [dependencies]
 serde = { version = "^1.0", features = ["derive"]}
 
-serde-wasm-bindgen = { version = "^0.5"}
+serde-wasm-bindgen = { version = "^0.6"}
 console_error_panic_hook = { version = "0.1.7"}
 wasm-bindgen = { version = "^0.2"}
 
 num_enum = "0.7.0"
-lol_alloc = "0.4.0"
+lol_alloc = { version = "0.4.0", default-features = false}
 piecewise-linear = "0.1.0"
+thiserror = "1.0.48"
 
 
 [lib]
@@ -34,4 +37,5 @@ crate-type = ["cdylib"]
 
 [profile.release]
 lto = true
+strip = true
 opt-level = "z"
