+++ src/activity/mod.rs
@@ -1,9 +1,8 @@
-use self::damage_calc::{get_gear_delta_mult, rpl_mult, DifficultyOptions, get_wep_delta_mult};
+use self::damage_calc::{get_gear_delta_mult, get_wep_delta_mult, rpl_mult, DifficultyOptions};
 
 pub mod damage_calc;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[derive(Default)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
 pub enum PlayerClass {
     #[default]
     Unknown = 0,
@@ -11,7 +10,6 @@ pub enum PlayerClass {
     Hunter = 2,
     Warlock = 3,
 }
-
 
 #[derive(Debug, Clone, Default)]
 pub struct Player {
@@ -46,7 +44,7 @@ impl Default for Activity {
 }
 impl Activity {
     pub fn get_pl_delta(&self) -> f64 {
-        get_gear_delta_mult(self)*get_wep_delta_mult(self)
+        get_gear_delta_mult(self) * get_wep_delta_mult(self)
     }
     pub fn get_rpl_mult(&self) -> f64 {
         rpl_mult(self.rpl as f64)
