+++ src/enemies/mod.rs
@@ -1,7 +1,6 @@
 use crate::activity::Activity;
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[derive(Default)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
 pub enum EnemyType {
     MINOR,
     ELITE,
@@ -14,7 +13,6 @@ pub enum EnemyType {
     CHAMPION,
 }
 
-
 #[derive(Debug, Clone, Default)]
 pub struct Enemy {
     pub health: f64,
@@ -26,5 +24,20 @@ pub struct Enemy {
 impl Enemy {
     pub fn get_adjusted_health(&self, _activity: Activity) -> f64 {
         self.health * (1.0 - self.damage_resistance)
+    }
+}
+
+//is here to sanitize the formulas_types.rs file for build script
+impl crate::types::formula_types::DamageModFormula {
+    pub fn get_mod(&self, _type: &EnemyType) -> f64 {
+        match *_type {
+            EnemyType::MINOR => self.minor,
+            EnemyType::ELITE => self.elite,
+            EnemyType::MINIBOSS => self.miniboss,
+            EnemyType::CHAMPION => self.champion,
+            EnemyType::BOSS => self.boss,
+            EnemyType::VEHICLE => self.vehicle,
+            _ => 1.0,
+        }
     }
 }
