+++ src/lib.rs
@@ -1,5 +1,3 @@
-#![allow(dead_code)]
-#![allow(unused_imports)]
 extern crate alloc;
 
 use logging::LogLevel;
@@ -15,24 +13,25 @@ mod test;
 pub mod types;
 pub mod weapons;
 
-use crate::perks::{Perk, Perks};
+use crate::perks::Perk;
 use crate::weapons::{Stat, Weapon};
 use abilities::Ability;
 use activity::Activity;
-use d2_enums::StatHashes;
 use enemies::Enemy;
 use std::cell::RefCell;
 use std::collections::HashMap;
 use std::panic;
 
-#[cfg(target_arch = "wasm32")]
-use lol_alloc::{AssumeSingleThreaded, FreeListAllocator};
 
 // SAFETY: This application is single threaded, so using AssumeSingleThreaded is allowed.
+#[cfg(target_arch = "wasm32")]
+use lol_alloc::{AssumeSingleThreaded, FreeListAllocator};
 #[cfg(target_arch = "wasm32")]
 #[global_allocator]
 static ALLOCATOR: AssumeSingleThreaded<FreeListAllocator> =
     unsafe { AssumeSingleThreaded::new(FreeListAllocator::new()) };
+
+
 mod built_info {
     include!(concat!(env!("OUT_DIR"), "/built.rs"));
 }
@@ -43,11 +42,7 @@ mod database {
 
 //JavaScript
 
-use crate::types::js_types::{
-    JsAmmoResponse, JsDifficultyOptions, JsDpsResponse, JsEnemyType, JsFiringResponse,
-    JsHandlingResponse, JsMetaData, JsRangeResponse, JsReloadResponse, JsResillienceSummary,
-    JsStat,
-};
+use crate::types::prelude::*;
 
 use wasm_bindgen::prelude::*;
