+++ src/perks/buff_perks.rs
@@ -62,19 +62,19 @@ fn gbl_debuff(_cached_data: &mut HashMap<String, f64>, _desired_buff: f64) -> f6
 pub fn buff_perks() {
     add_dmr(
         Perks::WellOfRadiance,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = emp_buff(_input.cached_data, 1.25);
             DamageModifierResponse {
                 impact_dmg_scale: buff,
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::NobleRounds,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value == 0 {
                 return DamageModifierResponse::default();
             }
@@ -85,12 +85,12 @@ pub fn buff_perks() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::Radiant,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let des_buff = if _input.pvp { 1.1 } else { 1.25 };
             let buff = emp_buff(_input.cached_data, des_buff);
             _input.cached_data.insert("radiant".to_string(), 1.0);
@@ -99,24 +99,24 @@ pub fn buff_perks() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::PathOfTheBurningSteps,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = surge_buff(_input.cached_data, _input.value, _input.pvp);
             DamageModifierResponse {
                 impact_dmg_scale: buff,
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::BannerShield,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let des_buff = if _input.pvp { 1.35 } else { 1.4 };
             let buff = emp_buff(_input.cached_data, des_buff);
             DamageModifierResponse {
@@ -124,12 +124,12 @@ pub fn buff_perks() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::EmpRift,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let des_buff = if _input.pvp { 1.15 } else { 1.2 };
             let buff = emp_buff(_input.cached_data, des_buff);
             DamageModifierResponse {
@@ -137,24 +137,24 @@ pub fn buff_perks() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::WardOfDawn,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = emp_buff(_input.cached_data, 1.25);
             DamageModifierResponse {
                 impact_dmg_scale: buff,
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::Gyrfalcon,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let des_buff = if _input.pvp { 1.0 } else { 1.35 };
             let buff = emp_buff(_input.cached_data, des_buff);
             DamageModifierResponse {
@@ -162,12 +162,12 @@ pub fn buff_perks() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::AeonInsight,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value > 0 {
                 let des_buff = if _input.pvp { 1.0 } else { 1.35 };
                 let buff = emp_buff(_input.cached_data, des_buff);
@@ -179,12 +179,12 @@ pub fn buff_perks() {
             } else {
                 DamageModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::UmbralSharpening,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let pve_values = [1.2, 1.25, 1.35, 1.4];
             let des_buff = if _input.pvp {
                 1.0
@@ -197,12 +197,12 @@ pub fn buff_perks() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::WormByproduct,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value > 0 {
                 DamageModifierResponse {
                     impact_dmg_scale: 1.15,
@@ -212,7 +212,7 @@ pub fn buff_perks() {
             } else {
                 DamageModifierResponse::default()
             }
-        }),
+        },
     );
 
     //
@@ -221,7 +221,7 @@ pub fn buff_perks() {
 
     add_dmr(
         Perks::Weaken,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let des_debuff = if _input.pvp { 1.075 } else { 1.15 };
             let debuff = gbl_debuff(_input.cached_data, des_debuff);
             DamageModifierResponse {
@@ -229,12 +229,12 @@ pub fn buff_perks() {
                 explosive_dmg_scale: debuff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::TractorCannon,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let des_debuff = if _input.pvp { 1.5 } else { 1.3 };
             let debuff = gbl_debuff(_input.cached_data, des_debuff);
             DamageModifierResponse {
@@ -242,12 +242,12 @@ pub fn buff_perks() {
                 explosive_dmg_scale: debuff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::MoebiusQuiver,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let des_debuff = if _input.pvp { 1.5 } else { 1.3 };
             let debuff = gbl_debuff(_input.cached_data, des_debuff);
             DamageModifierResponse {
@@ -255,11 +255,11 @@ pub fn buff_perks() {
                 explosive_dmg_scale: debuff,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_dmr(
         Perks::DeadFall,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let des_debuff = if _input.pvp { 1.5 } else { 1.3 };
             let debuff = gbl_debuff(_input.cached_data, des_debuff);
             DamageModifierResponse {
@@ -267,11 +267,11 @@ pub fn buff_perks() {
                 explosive_dmg_scale: debuff,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_dmr(
         Perks::Felwinters,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value > 0 {
                 let debuff = gbl_debuff(_input.cached_data, 1.3);
                 DamageModifierResponse {
@@ -282,12 +282,12 @@ pub fn buff_perks() {
             } else {
                 DamageModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::EnhancedScannerAugment,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let pve_values = [1.08, 1.137, 1.173, 1.193, 1.2];
             let des_debuff = if _input.pvp {
                 1.0
@@ -300,34 +300,34 @@ pub fn buff_perks() {
                 explosive_dmg_scale: debuff,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_dmr(
         Perks::SurgeMod,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let damage_mod = surge_buff(_input.cached_data, _input.value, _input.pvp);
             DamageModifierResponse {
                 explosive_dmg_scale: damage_mod,
                 impact_dmg_scale: damage_mod,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_dmr(
         Perks::EternalWarrior,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let damage_mod = surge_buff(_input.cached_data, _input.value, _input.pvp);
             DamageModifierResponse {
                 explosive_dmg_scale: damage_mod,
                 impact_dmg_scale: damage_mod,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::MantleOfBattleHarmony,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = if _input.value > 0 {
                 surge_buff(_input.cached_data, 4, _input.pvp)
             } else {
@@ -338,11 +338,11 @@ pub fn buff_perks() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_dmr(
         Perks::MaskOfBakris,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = if _input.value > 0
                 && matches!(
                     _input.calc_data.damage_type,
@@ -357,11 +357,11 @@ pub fn buff_perks() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_dmr(
         Perks::SanguineAlchemy,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value == 0 || *_input.calc_data.damage_type == DamageType::KINETIC {
                 return DamageModifierResponse::default();
             }
@@ -373,11 +373,11 @@ pub fn buff_perks() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_dmr(
         Perks::Foetracers,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value == 0 {
                 return DamageModifierResponse::default();
             }
@@ -387,11 +387,11 @@ pub fn buff_perks() {
                 explosive_dmg_scale: mult,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_dmr(
         Perks::GlacialGuard,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value == 0 {
                 return DamageModifierResponse::default();
             }
@@ -401,6 +401,6 @@ pub fn buff_perks() {
                 explosive_dmg_scale: mult,
                 ..Default::default()
             }
-        }),
+        },
     );
 }
