+++ src/perks/exotic_armor.rs
@@ -19,7 +19,7 @@ use super::{
 pub fn exotic_armor() {
     add_dmr(
         Perks::BallindorseWrathweavers,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut modifier = DamageModifierResponse::default();
             let value = if _input.pvp { 1.05 } else { 1.15 };
             if _input.calc_data.damage_type == &DamageType::STASIS && _input.value >= 1 {
@@ -27,12 +27,12 @@ pub fn exotic_armor() {
                 modifier.explosive_dmg_scale = value;
             }
             modifier
-        }),
+        },
     );
 
     add_dmr(
         Perks::MechaneersTricksleeves,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value == 0 || *_input.calc_data.weapon_type != WeaponType::SIDEARM {
                 return DamageModifierResponse::default();
             };
@@ -43,13 +43,13 @@ pub fn exotic_armor() {
                 impact_dmg_scale: damage_mult,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     //doesnt work for sturm overcharge, (maybe) memento
     add_dmr(
         Perks::LuckyPants,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let perks = _input.calc_data.perk_value_map.clone();
 
             let perk_check =
@@ -73,45 +73,41 @@ pub fn exotic_armor() {
                 impact_dmg_scale: 1.0 + mult * _input.value.clamp(0, 10) as f64,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::TomeOfDawn,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                if _input.value > 0 {
-                    stats.insert(StatHashes::AIRBORNE.into(), 50);
-                }
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            if _input.value > 0 {
+                stats.insert(StatHashes::AIRBORNE.into(), 50);
+            }
+            stats
+        },
     );
 
     add_flmr(
         Perks::TomeOfDawn,
-        Box::new(|_input: ModifierResponseInput| -> FlinchModifierResponse {
+        |_input: ModifierResponseInput| -> FlinchModifierResponse {
             if _input.value > 0 {
                 FlinchModifierResponse { flinch_scale: 0.80 }
             } else {
                 FlinchModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::KnuckleheadRadar,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                HashMap::from([(StatHashes::AIRBORNE.into(), 20)])
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            HashMap::from([(StatHashes::AIRBORNE.into(), 20)])
+        },
     );
 
     add_dmr(
         Perks::KnuckleheadRadar,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let health_percent = *_input.cached_data.get("health%").unwrap_or(&1.0);
             if health_percent >= 0.3 || _input.value == 0 {
                 return DamageModifierResponse::default();
@@ -122,44 +118,40 @@ pub fn exotic_armor() {
                 explosive_dmg_scale: modifier,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     //TODO: MECHANEER'S TRICKSLEEVES AUTORELOAD
 
     add_sbr(
         Perks::MechaneersTricksleeves,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                if _input.calc_data.weapon_type == &WeaponType::SIDEARM {
-                    stats.insert(StatHashes::AIRBORNE.into(), 50);
-                    stats.insert(StatHashes::HANDLING.into(), 100);
-                    stats.insert(StatHashes::RELOAD.into(), 100);
-                };
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            if _input.calc_data.weapon_type == &WeaponType::SIDEARM {
+                stats.insert(StatHashes::AIRBORNE.into(), 50);
+                stats.insert(StatHashes::HANDLING.into(), 100);
+                stats.insert(StatHashes::RELOAD.into(), 100);
+            };
+            stats
+        },
     );
 
     add_hmr(
         Perks::MechaneersTricksleeves,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.calc_data.weapon_type == &WeaponType::SIDEARM {
-                    HandlingModifierResponse {
-                        stat_add: 100,
-                        ..Default::default()
-                    }
-                } else {
-                    HandlingModifierResponse::default()
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.calc_data.weapon_type == &WeaponType::SIDEARM {
+                HandlingModifierResponse {
+                    stat_add: 100,
+                    ..Default::default()
                 }
-            },
-        ),
+            } else {
+                HandlingModifierResponse::default()
+            }
+        },
     );
     add_rsmr(
         Perks::MechaneersTricksleeves,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             if _input.calc_data.weapon_type == &WeaponType::SIDEARM {
                 ReloadModifierResponse {
                     reload_stat_add: 100,
@@ -168,12 +160,12 @@ pub fn exotic_armor() {
             } else {
                 ReloadModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::MechaneersTricksleeves,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mult = if _input.pvp { 1.35 } else { 2.0 };
             if _input.value > 0 && _input.calc_data.weapon_type == &WeaponType::SIDEARM {
                 DamageModifierResponse {
@@ -184,26 +176,24 @@ pub fn exotic_armor() {
             } else {
                 DamageModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::Oathkeeper,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                if _input.calc_data.weapon_type == &WeaponType::BOW {
-                    stats.insert(StatHashes::AIRBORNE.into(), 40);
-                    stats.insert(StatHashes::DRAW_TIME.into(), 10);
-                };
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            if _input.calc_data.weapon_type == &WeaponType::BOW {
+                stats.insert(StatHashes::AIRBORNE.into(), 40);
+                stats.insert(StatHashes::DRAW_TIME.into(), 10);
+            };
+            stats
+        },
     );
 
     /*add_fmr(
         Perks::Oathkeeper,
-        Box::new(|_input: ModifierResponsInput| -> FiringModifierResponse {
+        |_input: ModifierResponsInput| -> FiringModifierResponse {
             FiringModifierResponse {
                 burst_delay_add: match _input.calc_data.intrinsic_hash {
                     906 => -36.0 / 1100.0,
@@ -212,20 +202,18 @@ pub fn exotic_armor() {
                 },
                 ..Default::default()
             }
-        }),
+        },
     );*/
 
     add_sbr(
         Perks::SealedAhamkaraGrasps,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                if _input.value > 0 {
-                    stats.insert(StatHashes::AIRBORNE.into(), 50);
-                };
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            if _input.value > 0 {
+                stats.insert(StatHashes::AIRBORNE.into(), 50);
+            };
+            stats
+        },
     );
 
     //TODO: AUTORELOAD FOR SEALED AHAMKARA GRASPS
@@ -233,50 +221,44 @@ pub fn exotic_armor() {
     //LUCKY PANTS ONLY WORKS FOR READY ?!?!?! crazy :(
     add_sbr(
         Perks::LuckyPants,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stat = HashMap::new();
-                if _input.value > 0 && _input.calc_data.weapon_type == &WeaponType::HANDCANNON {
-                    stat.insert(StatHashes::AIRBORNE.into(), 20);
-                    stat.insert(StatHashes::HANDLING.into(), 100);
-                };
-                stat
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stat = HashMap::new();
+            if _input.value > 0 && _input.calc_data.weapon_type == &WeaponType::HANDCANNON {
+                stat.insert(StatHashes::AIRBORNE.into(), 20);
+                stat.insert(StatHashes::HANDLING.into(), 100);
+            };
+            stat
+        },
     );
 
     add_hmr(
         Perks::LuckyPants,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.value > 0 && _input.calc_data.weapon_type == &WeaponType::HANDCANNON {
-                    return HandlingModifierResponse {
-                        draw_add: 100,
-                        draw_scale: 0.6,
-                        ..Default::default()
-                    };
-                }
-                HandlingModifierResponse::default()
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.value > 0 && _input.calc_data.weapon_type == &WeaponType::HANDCANNON {
+                return HandlingModifierResponse {
+                    draw_add: 100,
+                    draw_scale: 0.6,
+                    ..Default::default()
+                };
+            }
+            HandlingModifierResponse::default()
+        },
     );
 
     add_sbr(
         Perks::NoBackupPlans,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                if _input.calc_data.weapon_type == &WeaponType::SHOTGUN {
-                    stats.insert(StatHashes::AIRBORNE.into(), 30);
-                };
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            if _input.calc_data.weapon_type == &WeaponType::SHOTGUN {
+                stats.insert(StatHashes::AIRBORNE.into(), 30);
+            };
+            stats
+        },
     );
 
     add_dmr(
         Perks::NoBackupPlans,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if *_input.calc_data.weapon_type != WeaponType::SHOTGUN || _input.value == 0 {
                 return DamageModifierResponse::default();
             }
@@ -287,220 +269,192 @@ pub fn exotic_armor() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::ActiumWarRig,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                if _input.calc_data.weapon_type == &WeaponType::AUTORIFLE
-                    || _input.calc_data.weapon_type == &WeaponType::MACHINEGUN
-                {
-                    stats.insert(StatHashes::AIRBORNE.into(), 30);
-                }
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            if _input.calc_data.weapon_type == &WeaponType::AUTORIFLE
+                || _input.calc_data.weapon_type == &WeaponType::MACHINEGUN
+            {
+                stats.insert(StatHashes::AIRBORNE.into(), 30);
+            }
+            stats
+        },
     );
 
     //TODO: AUTORELOAD ON ACTIUM WAR RIG
 
     add_sbr(
         Perks::HallowfireHeart,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                HashMap::from([(StatHashes::AIRBORNE.into(), 20)])
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            HashMap::from([(StatHashes::AIRBORNE.into(), 20)])
+        },
     );
 
     add_sbr(
         Perks::LionRampart,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                if _input.value > 0 {
-                    stats.insert(StatHashes::AIRBORNE.into(), 50);
-                };
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            if _input.value > 0 {
+                stats.insert(StatHashes::AIRBORNE.into(), 50);
+            };
+            stats
+        },
     );
 
     add_sbr(
         Perks::Peacekeepers,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                if _input.calc_data.weapon_type == &WeaponType::SUBMACHINEGUN {
-                    stats.insert(StatHashes::AIRBORNE.into(), 40);
-                    stats.insert(StatHashes::HANDLING.into(), 100);
-                };
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            if _input.calc_data.weapon_type == &WeaponType::SUBMACHINEGUN {
+                stats.insert(StatHashes::AIRBORNE.into(), 40);
+                stats.insert(StatHashes::HANDLING.into(), 100);
+            };
+            stats
+        },
     );
 
     add_hmr(
         Perks::Peacekeepers,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.calc_data.weapon_type == &WeaponType::SUBMACHINEGUN {
-                    return HandlingModifierResponse {
-                        stat_add: 100,
-                        ads_scale: 1.0,
-                        draw_scale: 0.6,
-                        stow_scale: 0.6,
-                        ..Default::default()
-                    };
-                }
-                HandlingModifierResponse::default()
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.calc_data.weapon_type == &WeaponType::SUBMACHINEGUN {
+                return HandlingModifierResponse {
+                    stat_add: 100,
+                    ads_scale: 1.0,
+                    draw_scale: 0.6,
+                    stow_scale: 0.6,
+                    ..Default::default()
+                };
+            }
+            HandlingModifierResponse::default()
+        },
     );
 
     add_sbr(
         Perks::PeregrineGreaves,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                HashMap::from([(StatHashes::AIRBORNE.into(), 20)])
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            HashMap::from([(StatHashes::AIRBORNE.into(), 20)])
+        },
     );
 
     add_sbr(
         Perks::EyeOfAnotherWorld,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                HashMap::from([(StatHashes::AIRBORNE.into(), 15)])
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            HashMap::from([(StatHashes::AIRBORNE.into(), 15)])
+        },
     );
 
     add_sbr(
         Perks::AstrocyteVerse,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                stats.insert(StatHashes::AIRBORNE.into(), 30);
-                if _input.value > 0 {
-                    stats.insert(StatHashes::HANDLING.into(), 100);
-                }
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            stats.insert(StatHashes::AIRBORNE.into(), 30);
+            if _input.value > 0 {
+                stats.insert(StatHashes::HANDLING.into(), 100);
+            }
+            stats
+        },
     );
 
     add_hmr(
         Perks::AstrocyteVerse,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.value == 0 {
-                    return HandlingModifierResponse::default();
-                }
-                HandlingModifierResponse {
-                    draw_add: 100,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.value == 0 {
+                return HandlingModifierResponse::default();
+            }
+            HandlingModifierResponse {
+                draw_add: 100,
+                ..Default::default()
+            }
+        },
     );
 
     add_sbr(
         Perks::NecroticGrips,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                if _input.calc_data.intrinsic_hash == 1863355414
-                    || _input.calc_data.intrinsic_hash == 2965975126
-                    || _input.calc_data.intrinsic_hash == 2724693746
-                {
-                    //Thorn, Osteo Striga, Touch of Malice
-                    stats.insert(StatHashes::AIRBORNE.into(), 30);
-                };
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            if _input.calc_data.intrinsic_hash == 1863355414
+                || _input.calc_data.intrinsic_hash == 2965975126
+                || _input.calc_data.intrinsic_hash == 2724693746
+            {
+                //Thorn, Osteo Striga, Touch of Malice
+                stats.insert(StatHashes::AIRBORNE.into(), 30);
+            };
+            stats
+        },
     );
 
     add_sbr(
         Perks::BootsOfTheAssembler,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                if _input.calc_data.intrinsic_hash == 2144092201 {
-                    //Lumina
-                    stats.insert(StatHashes::AIRBORNE.into(), 30);
-                };
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            if _input.calc_data.intrinsic_hash == 2144092201 {
+                //Lumina
+                stats.insert(StatHashes::AIRBORNE.into(), 30);
+            };
+            stats
+        },
     );
 
     add_sbr(
         Perks::RainOfFire,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stats = HashMap::new();
-                if _input.calc_data.weapon_type == &WeaponType::FUSIONRIFLE
-                    || _input.calc_data.weapon_type == &WeaponType::LINEARFUSIONRIFLE
-                {
-                    stats.insert(StatHashes::AIRBORNE.into(), 30);
-                }
-                stats
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stats = HashMap::new();
+            if _input.calc_data.weapon_type == &WeaponType::FUSIONRIFLE
+                || _input.calc_data.weapon_type == &WeaponType::LINEARFUSIONRIFLE
+            {
+                stats.insert(StatHashes::AIRBORNE.into(), 30);
+            }
+            stats
+        },
     );
 
     add_sbr(
         Perks::SpeedloaderSlacks,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let modifiers = match _input.value {
-                    0 => (0, 0, 0),
-                    1 => (40, 40, 30),
-                    2 => (40, 40, 35),
-                    3 => (45, 45, 40),
-                    4 => (50, 50, 45),
-                    5 => (55, 55, 50),
-                    _ => (55, 55, 50),
-                };
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let modifiers = match _input.value {
+                0 => (0, 0, 0),
+                1 => (40, 40, 30),
+                2 => (40, 40, 35),
+                3 => (45, 45, 40),
+                4 => (50, 50, 45),
+                5 => (55, 55, 50),
+                _ => (55, 55, 50),
+            };
 
-                HashMap::from([
-                    (StatHashes::RELOAD.into(), modifiers.0),
-                    (StatHashes::HANDLING.into(), modifiers.1), //?
-                    (StatHashes::AIRBORNE.into(), modifiers.2),
-                ])
-            },
-        ),
+            HashMap::from([
+                (StatHashes::RELOAD.into(), modifiers.0),
+                (StatHashes::HANDLING.into(), modifiers.1), //?
+                (StatHashes::AIRBORNE.into(), modifiers.2),
+            ])
+        },
     );
 
     add_hmr(
         Perks::SpeedloaderSlacks,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let handling = match _input.value {
-                    0 => 0,
-                    1 => 40,
-                    2 => 40,
-                    3 => 45,
-                    4 => 50,
-                    5 => 55,
-                    _ => 55,
-                };
-                HandlingModifierResponse {
-                    stat_add: handling,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let handling = match _input.value {
+                0 => 0,
+                1 => 40,
+                2 => 40,
+                3 => 45,
+                4 => 50,
+                5 => 55,
+                _ => 55,
+            };
+            HandlingModifierResponse {
+                stat_add: handling,
+                ..Default::default()
+            }
+        },
     );
 
     add_rsmr(
         Perks::SpeedloaderSlacks,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let modifiers = match _input.value {
                 0 => (0, 1.0),
                 1 => (40, 1.0),
@@ -515,25 +469,23 @@ pub fn exotic_armor() {
                 reload_stat_add: modifiers.0,
                 reload_time_scale: modifiers.1,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::LunaFaction,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let mut stat = HashMap::new();
-                if _input.value >= 1 {
-                    stat.insert(StatHashes::RELOAD.into(), 100);
-                }
-                stat
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let mut stat = HashMap::new();
+            if _input.value >= 1 {
+                stat.insert(StatHashes::RELOAD.into(), 100);
+            }
+            stat
+        },
     );
 
     add_rsmr(
         Perks::LunaFaction,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             if _input.value >= 1 {
                 ReloadModifierResponse {
                     reload_stat_add: 100,
@@ -542,12 +494,12 @@ pub fn exotic_armor() {
             } else {
                 ReloadModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_rmr(
         Perks::LunaFaction,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             if _input.value >= 2 {
                 return RangeModifierResponse {
                     range_all_scale: 2.0,
@@ -555,16 +507,16 @@ pub fn exotic_armor() {
                 };
             }
             RangeModifierResponse::default()
-        }),
+        },
     );
     add_sbr(
         Perks::TritonVice,
-        Box::new(|_input| -> HashMap<BungieHash, StatBump> {
+        |_input| -> HashMap<BungieHash, StatBump> {
             let mut stats = HashMap::new();
             if _input.value > 0 && *_input.calc_data.weapon_type == WeaponType::GLAIVE {
                 stats.insert(StatHashes::RELOAD.into(), 50);
             }
             stats
-        }),
-    )
+        },
+    );
 }
