+++ src/perks/exotic_perks.rs
@@ -1,8 +1,6 @@
 //This also includes intrinsic perks, not just exotic
 use std::collections::HashMap;
 
-use serde::__private::de;
-
 use crate::{d2_enums::StatHashes, enemies::EnemyType, weapons::Stat};
 
 use super::{
@@ -19,7 +17,7 @@ use super::{
 pub fn exotic_perks() {
     add_dmr(
         Perks::ParacausalShot,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let bufflist_pve = vec![1.0, 3.92, 4.0, 4.4, 5.25, 7.67, 11.71, 18.36];
             let bufflist_pvp = vec![1.0, 1.01, 1.03, 1.13, 1.41, 1.96, 3.0, 4.73];
             let mut damage_buff = 1.0;
@@ -46,12 +44,12 @@ pub fn exotic_perks() {
                 explosive_dmg_scale: damage_buff,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::HuntersTrance,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut out = HashMap::new();
             let inter_val = *_input
                 .calc_data
@@ -63,12 +61,12 @@ pub fn exotic_perks() {
             out.insert(StatHashes::RANGE.into(), buff_val);
             out.insert(StatHashes::HANDLING.into(), buff_val);
             out
-        }),
+        },
     );
 
     add_rsmr(
         Perks::HuntersTrance,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let inter_val = *_input
                 .calc_data
                 .perk_value_map
@@ -79,12 +77,12 @@ pub fn exotic_perks() {
                 reload_stat_add: buff_val,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_rmr(
         Perks::HuntersTrance,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let inter_val = *_input
                 .calc_data
                 .perk_value_map
@@ -95,41 +93,39 @@ pub fn exotic_perks() {
                 range_stat_add: buff_val,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_hmr(
         Perks::HuntersTrance,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let inter_val = *_input
-                    .calc_data
-                    .perk_value_map
-                    .get(&213689231)
-                    .unwrap_or(&0);
-                let buff_val = (clamp(inter_val, 0, 7) * 5) as i32;
-                HandlingModifierResponse {
-                    stat_add: buff_val,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let inter_val = *_input
+                .calc_data
+                .perk_value_map
+                .get(&213689231)
+                .unwrap_or(&0);
+            let buff_val = (clamp(inter_val, 0, 7) * 5) as i32;
+            HandlingModifierResponse {
+                stat_add: buff_val,
+                ..Default::default()
+            }
+        },
     );
 
     add_rmr(
         Perks::HuntersTrace,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let range_ads_scale = if _input.value > 0 { 4.5 / 1.7 } else { 1.0 };
             RangeModifierResponse {
                 range_zoom_scale: range_ads_scale,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::MementoMori,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut damage_buff = 1.0;
             if _input.value > 0 && _input.calc_data.total_shots_fired < 7.0 {
                 damage_buff = if _input.pvp { 1.285 } else { 1.5 };
@@ -139,12 +135,12 @@ pub fn exotic_perks() {
                 explosive_dmg_scale: damage_buff,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_rmr(
         Perks::MementoMori,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let range_all_scale = if _input.value > 0 && _input.calc_data.total_shots_fired < 7.0 {
                 0.85
             } else {
@@ -154,24 +150,24 @@ pub fn exotic_perks() {
                 range_all_scale,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::Roadborn,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut out = HashMap::new();
             if _input.value > 0 {
                 out.insert(StatHashes::HANDLING.into(), 20);
                 out.insert(StatHashes::RELOAD.into(), 40);
             };
             out
-        }),
+        },
     );
 
     add_dmr(
         Perks::Roadborn,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut crit_mult = 1.0;
             if _input.value > 0 {
                 crit_mult = 1.17;
@@ -181,12 +177,12 @@ pub fn exotic_perks() {
                 explosive_dmg_scale: 1.0,
                 impact_dmg_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::Roadborn,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             let mut delay_mult = 1.0;
             if _input.value > 0 {
                 delay_mult = 0.583;
@@ -197,12 +193,12 @@ pub fn exotic_perks() {
                 inner_burst_scale: 1.0,
                 burst_size_add: 0.0,
             }
-        }),
+        },
     );
 
     add_rmr(
         Perks::Roadborn,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let mut range_scale = 1.05;
             if _input.value > 0 {
                 range_scale = 1.15; //roughly
@@ -213,12 +209,12 @@ pub fn exotic_perks() {
                 range_hip_scale: 1.0,
                 range_zoom_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_rsmr(
         Perks::Roadborn,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let mut reload = 0;
             if _input.value > 0 {
                 reload = 40;
@@ -227,12 +223,12 @@ pub fn exotic_perks() {
                 reload_stat_add: reload,
                 reload_time_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::ReignHavoc,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             let mut delay_mult = 1.0;
             if _input.calc_data.shots_fired_this_mag >= _input.calc_data.base_mag * 0.2 {
                 delay_mult = 0.75;
@@ -246,12 +242,12 @@ pub fn exotic_perks() {
                 inner_burst_scale: 1.0,
                 burst_size_add: 0.0,
             }
-        }),
+        },
     );
 
     add_edr(
         Perks::ReignHavoc,
-        Box::new(|_input: ModifierResponseInput| -> ExtraDamageResponse {
+        |_input: ModifierResponseInput| -> ExtraDamageResponse {
             let dmg = if _input.pvp { 65.0 } else { 65.0 * 1.3 };
             ExtraDamageResponse {
                 additive_damage: dmg,
@@ -264,24 +260,24 @@ pub fn exotic_perks() {
                 crit_scale: false,
                 combatant_scale: true,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::WormsHunger,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let val = clamp(_input.value, 0, 20);
             DamageModifierResponse {
                 impact_dmg_scale: 1.0 + (val as f64) * 0.1,
                 explosive_dmg_scale: 1.0 + (val as f64) * 0.1,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::LagragianSight,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut damage_buff = 1.0;
             if _input.value > 0 && _input.calc_data.time_total < 30.0 {
                 damage_buff = 1.4;
@@ -291,12 +287,12 @@ pub fn exotic_perks() {
                 explosive_dmg_scale: damage_buff,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::ToM,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut damage_buff = 1.0;
             if _input.calc_data.curr_mag == 1.0 {
                 damage_buff = if _input.pvp { 2.0 } else { 2.4 };
@@ -306,12 +302,12 @@ pub fn exotic_perks() {
                 explosive_dmg_scale: damage_buff,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_rr(
         Perks::ToM,
-        Box::new(|_input: ModifierResponseInput| -> RefundResponse {
+        |_input: ModifierResponseInput| -> RefundResponse {
             RefundResponse {
                 refund_mag: if _input.calc_data.curr_mag == 0.0 {
                     1
@@ -322,12 +318,12 @@ pub fn exotic_perks() {
                 crit: false,
                 requirement: 1,
             }
-        }),
+        },
     );
 
     add_edr(
         Perks::RocketTracers,
-        Box::new(|_input: ModifierResponseInput| -> ExtraDamageResponse {
+        |_input: ModifierResponseInput| -> ExtraDamageResponse {
             let dmg = if _input.pvp { 24.0 } else { 105.0 };
             ExtraDamageResponse {
                 additive_damage: dmg,
@@ -340,7 +336,7 @@ pub fn exotic_perks() {
                 crit_scale: false,
                 combatant_scale: true,
             }
-        }),
+        },
     );
 
     // add_edr_guidance_ring(
@@ -395,29 +391,29 @@ pub fn exotic_perks() {
 
     add_fmr(
         Perks::HakkeHeavyBurst,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             FiringModifierResponse {
                 burst_size_add: -2.0,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::HakkeHeavyBurst,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let crit_scale = (1.5 + 5.0 / 51.0) / _input.calc_data.base_crit_mult;
             DamageModifierResponse {
                 explosive_dmg_scale: 1.48,
                 impact_dmg_scale: 1.48,
                 crit_scale,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::SwoopingTalons,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut dmg_mult = 1.0;
             if _input.value > 0 {
                 dmg_mult = 1.4;
@@ -429,11 +425,12 @@ pub fn exotic_perks() {
                 explosive_dmg_scale: dmg_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
+
     add_dmr(
         Perks::IgnitionTrigger,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut dmg_mult = 1.0;
             if _input.value > 0 || _input.calc_data.total_shots_fired > 20.0 {
                 dmg_mult = if _input.pvp { 1.55 } else { 1.99 };
@@ -443,12 +440,12 @@ pub fn exotic_perks() {
                 explosive_dmg_scale: dmg_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::CalculatedBalance,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut damage_mult = if _input.value > 0 { 0.2 } else { 0.0 };
             let duration = 5.0;
             if _input.calc_data.time_total > duration {
@@ -459,12 +456,12 @@ pub fn exotic_perks() {
                 explosive_dmg_scale: 1.0 + damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::RavenousBeast,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             if _input.value > 0 {
                 FiringModifierResponse {
                     burst_delay_scale: 0.8,
@@ -473,12 +470,12 @@ pub fn exotic_perks() {
             } else {
                 FiringModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::RavenousBeast,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut damage_mult = 1.0;
             let mut crit_mult = 1.0;
             if _input.value > 0 {
@@ -494,12 +491,12 @@ pub fn exotic_perks() {
                 explosive_dmg_scale: damage_mult,
                 crit_scale: crit_mult,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::ReleaseTheWolves,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let has_cat = _input.calc_data.perk_value_map.contains_key(&431220296);
             let mut out = HashMap::new();
             if has_cat {
@@ -510,12 +507,12 @@ pub fn exotic_perks() {
                 }
             }
             out
-        }),
+        },
     );
 
     add_rsmr(
         Perks::ReleaseTheWolves,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let has_cat = _input.calc_data.perk_value_map.contains_key(&431220296);
             if _input.value == 1 && has_cat {
                 ReloadModifierResponse {
@@ -525,12 +522,12 @@ pub fn exotic_perks() {
             } else {
                 ReloadModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::ReleaseTheWolves,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             if _input.value > 0 {
                 FiringModifierResponse {
                     burst_delay_scale: 0.4,
@@ -539,24 +536,24 @@ pub fn exotic_perks() {
             } else {
                 FiringModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::ReleaseTheWolves,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let damage_mult = if _input.value > 0 { 1.4 } else { 1.0 };
             DamageModifierResponse {
                 impact_dmg_scale: damage_mult,
                 explosive_dmg_scale: damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::Fundamentals,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut stats = HashMap::new();
             if _input.value == 1 {
                 stats.insert(StatHashes::STABILITY.into(), 20);
@@ -569,28 +566,26 @@ pub fn exotic_perks() {
                 stats.insert(StatHashes::HANDLING.into(), 25);
             };
             stats
-        }),
+        },
     );
 
     add_hmr(
         Perks::Fundamentals,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let mut handling = 0;
-                if _input.value == 3 {
-                    handling = 25;
-                }
-                HandlingModifierResponse {
-                    stat_add: handling,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let mut handling = 0;
+            if _input.value == 3 {
+                handling = 25;
+            }
+            HandlingModifierResponse {
+                stat_add: handling,
+                ..Default::default()
+            }
+        },
     );
 
     add_rsmr(
         Perks::Fundamentals,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let mut reload = 0;
             if _input.value == 2 {
                 reload = 35;
@@ -599,12 +594,12 @@ pub fn exotic_perks() {
                 reload_stat_add: reload,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_rmr(
         Perks::Fundamentals,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let mut range = 0;
             if _input.value == 3 {
                 range = 5;
@@ -613,23 +608,23 @@ pub fn exotic_perks() {
                 range_stat_add: range,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::ThinTheHerd,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut out = HashMap::new();
             if _input.value > 0 {
                 out.insert(StatHashes::RELOAD.into(), 70);
             }
             out
-        }),
+        },
     );
 
     add_rsmr(
         Perks::ThinTheHerd,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             if _input.value > 0 {
                 ReloadModifierResponse {
                     reload_stat_add: 70,
@@ -638,39 +633,37 @@ pub fn exotic_perks() {
             } else {
                 ReloadModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_hmr(
         Perks::Chimera,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.value > 0 {
-                    HandlingModifierResponse {
-                        stat_add: 100,
-                        ..Default::default()
-                    }
-                } else {
-                    HandlingModifierResponse::default()
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.value > 0 {
+                HandlingModifierResponse {
+                    stat_add: 100,
+                    ..Default::default()
                 }
-            },
-        ),
+            } else {
+                HandlingModifierResponse::default()
+            }
+        },
     );
 
     add_sbr(
         Perks::Chimera,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut out = HashMap::new();
             if _input.value > 0 {
                 out.insert(StatHashes::RELOAD.into(), 100);
             }
             out
-        }),
+        },
     );
 
     add_dmr(
         Perks::FirstGlance,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut damage_mult = 1.0;
             let mut crit_mult = 1.0;
             if _input.value > 0 {
@@ -685,12 +678,12 @@ pub fn exotic_perks() {
                 impact_dmg_scale: damage_mult,
                 crit_scale: crit_mult,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::FateOfAllFools,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut damage_mult = 1.0;
             let mut crit_mult = 1.0;
             if _input.value as f64 > _input.calc_data.total_shots_fired {
@@ -703,12 +696,12 @@ pub fn exotic_perks() {
                 impact_dmg_scale: damage_mult,
                 crit_scale: crit_mult,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::HonedEdge,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut damage_mult = 1.0;
             let has_cat = _input.calc_data.perk_value_map.contains_key(&529188544);
             if _input.value == 2 {
@@ -725,12 +718,12 @@ pub fn exotic_perks() {
                 impact_dmg_scale: damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::TakenPredator,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut damage_mult = 1.0;
             if _input.value == 1 || _input.value == 2 {
                 damage_mult = 1.25;
@@ -742,12 +735,12 @@ pub fn exotic_perks() {
                 impact_dmg_scale: damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::MarkovChain,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let val = clamp(_input.value, 0, 5);
             let damage_mult = (1.0 / 15.0) * val as f64 * if _input.pvp { 1.0 } else { 2.0 };
             DamageModifierResponse {
@@ -755,12 +748,12 @@ pub fn exotic_perks() {
                 impact_dmg_scale: 1.0 + damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::StringofCurses,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let val = clamp(_input.value, 0, 5);
             let mut damage_mult = 0.2 * val as f64;
             if _input.pvp {
@@ -775,12 +768,12 @@ pub fn exotic_perks() {
                 explosive_dmg_scale: 1.0 + damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::StormAndStress,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value == 0 {
                 return DamageModifierResponse::default();
             }
@@ -791,12 +784,12 @@ pub fn exotic_perks() {
                 impact_dmg_scale: damage_mult,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_rmr(
         Perks::DualSpeedReceiver,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             if _input.value == 0 {
                 return RangeModifierResponse::default();
             }
@@ -804,35 +797,35 @@ pub fn exotic_perks() {
                 range_stat_add: 30,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::DualSpeedReceiver,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut out = HashMap::new();
             if _input.value > 0 {
                 out.insert(StatHashes::ZOOM.into(), 3);
                 out.insert(StatHashes::RANGE.into(), 30);
             }
             out
-        }),
+        },
     );
 
     add_dmr(
         Perks::FullStop,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             DamageModifierResponse {
                 explosive_dmg_scale: 1.0,
                 impact_dmg_scale: 1.0,
                 crit_scale: if !_input.pvp { 2.9 } else { 1.0 },
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::RatPack,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             if _input.value == 0 {
                 return FiringModifierResponse::default();
             }
@@ -842,83 +835,81 @@ pub fn exotic_perks() {
                 burst_delay_add: val as f64 * (-0.625 / 30.0),
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_mmr(
         Perks::RatPack,
-        Box::new(
-            |_input: ModifierResponseInput| -> MagazineModifierResponse {
-                let val = clamp(_input.value - 1, 0, 4);
-                MagazineModifierResponse {
-                    magazine_add: val as f64 * if val == 4 { 2.25 } else { 2.0 },
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> MagazineModifierResponse {
+            let val = clamp(_input.value - 1, 0, 4);
+            MagazineModifierResponse {
+                magazine_add: val as f64 * if val == 4 { 2.25 } else { 2.0 },
+                ..Default::default()
+            }
+        },
     );
 
     add_fmr(
         Perks::RideTheBull,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             let extra_value = _input.calc_data.shots_fired_this_mag / 10.0;
             let val = clamp(_input.value + extra_value as u32, 0, 2);
             FiringModifierResponse {
                 burst_delay_add: val as f64 * (-0.25 / 30.0),
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::SpinningUp,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             let extra_value = _input.calc_data.shots_fired_this_mag / 12.0;
             let val = clamp(_input.value + extra_value as u32, 0, 2);
             FiringModifierResponse {
                 burst_delay_add: val as f64 * (-0.5 / 30.0),
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::CranialSpike,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut out = HashMap::new();
             let val = clamp(_input.value, 0, 5) as i32;
             out.insert(StatHashes::RANGE.into(), 8 * val);
             out.insert(StatHashes::AIM_ASSIST.into(), 4 * val);
             out
-        }),
+        },
     );
 
     add_rsmr(
         Perks::CranialSpike,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let val = clamp(_input.value, 0, 5) as i32;
             let rel = 0.97_f64.powi(val);
             ReloadModifierResponse {
                 reload_time_scale: rel,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_rmr(
         Perks::CranialSpike,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let val = clamp(_input.value, 0, 5) as i32;
             RangeModifierResponse {
                 range_stat_add: 8 * val,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::DarkForgedTrigger,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             if _input.value == 0 {
                 if _input
                     .calc_data
@@ -940,12 +931,12 @@ pub fn exotic_perks() {
             } else {
                 FiringModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::HarmonicLaser,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = match (_input.value, _input.pvp) {
                 (0, _) => 1.0,
                 (1, true) => 1.03,
@@ -957,12 +948,12 @@ pub fn exotic_perks() {
                 impact_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::AgersScepterCatalyst,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value > 0 {
                 return DamageModifierResponse {
                     impact_dmg_scale: 1.8,
@@ -970,61 +961,59 @@ pub fn exotic_perks() {
                 };
             }
             DamageModifierResponse::default()
-        }),
+        },
     );
 
     add_mmr(
         Perks::AgersScepterCatalyst,
-        Box::new(
-            |_input: ModifierResponseInput| -> MagazineModifierResponse {
-                let mag_buff = if _input.value > 0 && _input.calc_data.total_shots_fired == 0.0 {
-                    2.0
-                } else {
-                    1.0
-                };
-                MagazineModifierResponse {
-                    magazine_scale: mag_buff,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> MagazineModifierResponse {
+            let mag_buff = if _input.value > 0 && _input.calc_data.total_shots_fired == 0.0 {
+                2.0
+            } else {
+                1.0
+            };
+            MagazineModifierResponse {
+                magazine_scale: mag_buff,
+                ..Default::default()
+            }
+        },
     );
 
     add_dmr(
         Perks::ColdFusion,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = 0.0195 * clamp(_input.calc_data.total_shots_hit, 0.0, 41.0);
             DamageModifierResponse {
                 impact_dmg_scale: 1.0 + buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     //Queenbreaker's sights
     add_dmr(
         Perks::MarksmanSights,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             DamageModifierResponse {
                 impact_dmg_scale: 1.38,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::MarksmanSights,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             FiringModifierResponse {
                 burst_delay_add: (1800.0 / (60000.0 / 333.0)), // 300 + 333 = 633 ,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::Broadside,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = match _input.value {
                 0 => 1.0,
                 1 => 1.18,
@@ -1036,12 +1025,12 @@ pub fn exotic_perks() {
                 impact_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::TemporalUnlimiter,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             if _input.value > 0 {
                 return FiringModifierResponse {
                     burst_delay_add: 0.366,
@@ -1049,12 +1038,12 @@ pub fn exotic_perks() {
                 };
             }
             FiringModifierResponse::default()
-        }),
+        },
     );
 
     add_dmr(
         Perks::TemporalUnlimiter,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = if _input.pvp { 7.545 } else { 14.0 };
             if _input.value > 0 {
                 return DamageModifierResponse {
@@ -1064,24 +1053,22 @@ pub fn exotic_perks() {
                 };
             }
             DamageModifierResponse::default()
-        }),
+        },
     );
 
     add_mmr(
         Perks::FourthHorsemanCatalyst,
-        Box::new(
-            |_input: ModifierResponseInput| -> MagazineModifierResponse {
-                MagazineModifierResponse {
-                    magazine_add: 1.0,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> MagazineModifierResponse {
+            MagazineModifierResponse {
+                magazine_add: 1.0,
+                ..Default::default()
+            }
+        },
     );
 
     add_dmr(
         Perks::BlackHole,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = if _input.calc_data.total_shots_hit % 2.0 == 1.0 {
                 1.35
             } else {
@@ -1091,12 +1078,12 @@ pub fn exotic_perks() {
                 impact_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::Impetus,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value > 0 {
                 return DamageModifierResponse {
                     impact_dmg_scale: 1.5,
@@ -1104,22 +1091,22 @@ pub fn exotic_perks() {
                 };
             }
             DamageModifierResponse::default()
-        }),
+        },
     );
 
     add_fmr(
         Perks::MarksmanSights,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             FiringModifierResponse {
                 burst_delay_add: 0.333, // 300 + 333 = 633 ,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::Broadhead,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let broadhead_damage = if _input.pvp { 30.0 } else { 60.0 };
             let impact_damage = _input.calc_data.curr_firing_data.damage;
             let crit_mult = _input.calc_data.curr_firing_data.crit_mult;
@@ -1134,11 +1121,11 @@ pub fn exotic_perks() {
                 crit_scale,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_fmr(
         Perks::Desperation,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             let duration = 7.0;
             if _input.value == 0 || _input.calc_data.time_total > duration {
                 return FiringModifierResponse::default();
@@ -1147,6 +1134,6 @@ pub fn exotic_perks() {
                 burst_delay_scale: 0.8,
                 ..Default::default()
             }
-        }),
+        },
     );
 }
