+++ src/perks/origin_perks.rs
@@ -16,12 +16,11 @@ use super::{
 pub fn origin_perks() {
     add_rr(
         Perks::VeistStinger,
-        Box::new(|_input: ModifierResponseInput| -> RefundResponse {
+        |_input: ModifierResponseInput| -> RefundResponse {
             let data = _input.cached_data.get("veist_stinger");
             let last_proc = *data.unwrap_or(&1.0);
             let time_since_last_proc = _input.calc_data.time_total - last_proc;
             let max_refund = _input.calc_data.base_mag - _input.calc_data.curr_mag;
-            
 
             if _input.value == 0 || time_since_last_proc < 4.0 || max_refund == 0.0 {
                 return RefundResponse::default();
@@ -38,12 +37,12 @@ pub fn origin_perks() {
                 refund_mag: refund_amount,
                 refund_reserves: -final_refund_ammount,
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::VeistStinger,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             FiringModifierResponse {
                 burst_delay_scale: if _input.calc_data.weapon_type == &WeaponType::BOW
                     && _input.value > 0
@@ -54,46 +53,46 @@ pub fn origin_perks() {
                 },
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::HakkeBreach,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let damage_mult = if _input.value > 0 { 0.3 } else { 0.0 };
             DamageModifierResponse {
                 impact_dmg_scale: 1.0 + damage_mult,
                 explosive_dmg_scale: 1.0 + damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_rmr(
         Perks::Alacrity,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let range_add = if _input.value > 0 { 20 } else { 0 };
             RangeModifierResponse {
                 range_stat_add: range_add,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_rsmr(
         Perks::Alacrity,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let reload_add = if _input.value > 0 { 50 } else { 0 };
             ReloadModifierResponse {
                 reload_stat_add: reload_add,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::Alacrity,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             let range = if _input.value > 0 { 20 } else { 0 };
             let reload = if _input.value > 0 { 50 } else { 0 };
@@ -104,12 +103,12 @@ pub fn origin_perks() {
             map.insert(StatHashes::STABILITY.into(), stability);
             map.insert(StatHashes::AIM_ASSIST.into(), aim_assist);
             map
-        }),
+        },
     );
 
     add_sbr(
         Perks::Ambush,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             let range = if _input.is_enhanced { 30 } else { 20 };
             let handling = if _input.is_enhanced { 40 } else { 20 };
@@ -118,12 +117,12 @@ pub fn origin_perks() {
                 map.insert(StatHashes::HANDLING.into(), handling);
             }
             map
-        }),
+        },
     );
 
     add_rmr(
         Perks::Ambush,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let range_add = if _input.is_enhanced { 30 } else { 20 };
             if _input.calc_data.time_total < 2.0 && _input.value > 0 {
                 RangeModifierResponse {
@@ -133,29 +132,27 @@ pub fn origin_perks() {
             } else {
                 RangeModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_hmr(
         Perks::Ambush,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let handling_add = if _input.is_enhanced { 40 } else { 20 };
-                if _input.calc_data.time_total < 2.0 && _input.value > 0 {
-                    HandlingModifierResponse {
-                        stat_add: handling_add,
-                        ..Default::default()
-                    }
-                } else {
-                    HandlingModifierResponse::default()
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let handling_add = if _input.is_enhanced { 40 } else { 20 };
+            if _input.calc_data.time_total < 2.0 && _input.value > 0 {
+                HandlingModifierResponse {
+                    stat_add: handling_add,
+                    ..Default::default()
                 }
-            },
-        ),
+            } else {
+                HandlingModifierResponse::default()
+            }
+        },
     );
 
     add_dmr(
         Perks::Ambush,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value == 0 || _input.pvp {
                 return DamageModifierResponse::default();
             }
@@ -170,12 +167,12 @@ pub fn origin_perks() {
                 explosive_dmg_scale: damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::Ambush,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             FiringModifierResponse {
                 burst_delay_scale: if _input.calc_data.weapon_type == &WeaponType::BOW
                     && _input.value > 0
@@ -186,29 +183,27 @@ pub fn origin_perks() {
                 },
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_hmr(
         Perks::HotSwap,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let handling_add = if _input.is_enhanced { 60 } else { 30 };
-                if _input.value > 0 {
-                    HandlingModifierResponse {
-                        stat_add: handling_add,
-                        ..Default::default()
-                    }
-                } else {
-                    HandlingModifierResponse::default()
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let handling_add = if _input.is_enhanced { 60 } else { 30 };
+            if _input.value > 0 {
+                HandlingModifierResponse {
+                    stat_add: handling_add,
+                    ..Default::default()
                 }
-            },
-        ),
+            } else {
+                HandlingModifierResponse::default()
+            }
+        },
     );
 
     add_rsmr(
         Perks::FluidDynamics,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let reload_add = if _input.is_enhanced { 35 } else { 30 };
             if _input.calc_data.shots_fired_this_mag <= _input.calc_data.base_mag / 2.0 {
                 ReloadModifierResponse {
@@ -218,12 +213,12 @@ pub fn origin_perks() {
             } else {
                 ReloadModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::FluidDynamics,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             let reload = if _input.is_enhanced { 35 } else { 30 };
             let stability = if _input.is_enhanced { 25 } else { 20 };
@@ -234,12 +229,12 @@ pub fn origin_perks() {
                 map.insert(StatHashes::STABILITY.into(), stability);
             }
             map
-        }),
+        },
     );
 
     add_rsmr(
         Perks::QuietMoment,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             if _input.value > 0 {
                 ReloadModifierResponse {
                     reload_stat_add: 40,
@@ -248,23 +243,23 @@ pub fn origin_perks() {
             } else {
                 ReloadModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::QuietMoment,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             if _input.value > 0 {
                 map.insert(StatHashes::RELOAD.into(), 40);
             }
             map
-        }),
+        },
     );
 
     add_rsmr(
         Perks::BitterSpite,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let val = clamp(_input.value, 0, 5) as i32;
             let mult = match val {
                 0 => 1.0,
@@ -279,22 +274,22 @@ pub fn origin_perks() {
                 reload_stat_add: val * 10,
                 reload_time_scale: mult,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::BitterSpite,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             let val = clamp(_input.value, 0, 5) as i32;
             map.insert(StatHashes::RELOAD.into(), val * 10);
             map
-        }),
+        },
     );
 
     add_rmr(
         Perks::RightHook,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let range_add = if _input.is_enhanced { 20 } else { 10 };
             if _input.value > 0 {
                 RangeModifierResponse {
@@ -304,12 +299,12 @@ pub fn origin_perks() {
             } else {
                 RangeModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::RightHook,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             let stat_bump = if _input.is_enhanced { 20 } else { 10 };
             if _input.value > 0 {
@@ -317,69 +312,63 @@ pub fn origin_perks() {
                 map.insert(StatHashes::RANGE.into(), stat_bump);
             }
             map
-        }),
+        },
     );
 
     add_hmr(
         Perks::SearchParty,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.value > 0 {
-                    HandlingModifierResponse {
-                        ads_scale: 0.85,
-                        ..Default::default()
-                    }
-                } else {
-                    HandlingModifierResponse::default()
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.value > 0 {
+                HandlingModifierResponse {
+                    ads_scale: 0.85,
+                    ..Default::default()
                 }
-            },
-        ),
+            } else {
+                HandlingModifierResponse::default()
+            }
+        },
     );
 
     add_mmr(
         Perks::RunnethOver,
-        Box::new(
-            |_input: ModifierResponseInput| -> MagazineModifierResponse {
-                let val = clamp(_input.value, 0, 5) as f64;
-                MagazineModifierResponse {
-                    magazine_scale: 1.0 + val * 0.1,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> MagazineModifierResponse {
+            let val = clamp(_input.value, 0, 5) as f64;
+            MagazineModifierResponse {
+                magazine_scale: 1.0 + val * 0.1,
+                ..Default::default()
+            }
+        },
     );
 
     add_sbr(
         Perks::TexBalancedStock,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             if _input.value > 0 {
                 map.insert(StatHashes::HANDLING.into(), 20);
                 map.insert(StatHashes::RELOAD.into(), 20);
             }
             map
-        }),
+        },
     );
 
     add_hmr(
         Perks::TexBalancedStock,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.value > 0 {
-                    HandlingModifierResponse {
-                        stat_add: 50,
-                        ..Default::default()
-                    }
-                } else {
-                    HandlingModifierResponse::default()
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.value > 0 {
+                HandlingModifierResponse {
+                    stat_add: 50,
+                    ..Default::default()
                 }
-            },
-        ),
+            } else {
+                HandlingModifierResponse::default()
+            }
+        },
     );
 
     add_rsmr(
         Perks::TexBalancedStock,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             if _input.value > 0 {
                 ReloadModifierResponse {
                     reload_stat_add: 20,
@@ -388,50 +377,48 @@ pub fn origin_perks() {
             } else {
                 ReloadModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::SurosSynergy,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut out = HashMap::new();
             if _input.value > 0 {
                 out.insert(StatHashes::HANDLING.into(), 40);
             }
             out
-        }),
+        },
     );
 
     add_hmr(
         Perks::SurosSynergy,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.value > 0 {
-                    HandlingModifierResponse {
-                        stat_add: 40,
-                        ..Default::default()
-                    }
-                } else {
-                    HandlingModifierResponse::default()
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.value > 0 {
+                HandlingModifierResponse {
+                    stat_add: 40,
+                    ..Default::default()
                 }
-            },
-        ),
+            } else {
+                HandlingModifierResponse::default()
+            }
+        },
     );
 
     add_flmr(
         Perks::SurosSynergy,
-        Box::new(|_input: ModifierResponseInput| -> FlinchModifierResponse {
+        |_input: ModifierResponseInput| -> FlinchModifierResponse {
             if _input.value > 0 {
                 FlinchModifierResponse { flinch_scale: 0.80 }
             } else {
                 FlinchModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::HarmonicResonance,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut out = HashMap::new();
             if _input.value == 1 {
                 out.insert(StatHashes::HANDLING.into(), 10);
@@ -441,12 +428,12 @@ pub fn origin_perks() {
                 out.insert(StatHashes::HANDLING.into(), 20);
             }
             out
-        }),
+        },
     );
 
     add_rsmr(
         Perks::HarmonicResonance,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let stat_bump = if _input.value > 1 { 20 } else { 0 };
             if _input.value > 0 {
                 ReloadModifierResponse {
@@ -456,19 +443,17 @@ pub fn origin_perks() {
             } else {
                 ReloadModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_hmr(
         Perks::HarmonicResonance,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let stat_bump = 10 * clamp(_input.value, 0, 2);
-                HandlingModifierResponse {
-                    stat_add: stat_bump as i32,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let stat_bump = 10 * clamp(_input.value, 0, 2);
+            HandlingModifierResponse {
+                stat_add: stat_bump as i32,
+                ..Default::default()
+            }
+        },
     );
 }
