+++ src/perks/year_2_perks.rs
@@ -19,18 +19,18 @@ use super::{
 pub fn year_2_perks() {
     add_sbr(
         Perks::AirAssault,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut stats = HashMap::new();
             let ae_per_stack = if _input.is_enhanced { 35 } else { 30 };
             let ae = ae_per_stack * _input.value as i32;
             stats.insert(StatHashes::AIRBORNE.into(), ae);
             stats
-        }),
+        },
     );
 
     add_fmr(
         Perks::ArchersTempo,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             let delay = match (_input.value, _input.calc_data.intrinsic_hash) {
                 (0, _) => 1.0,
                 (1.., 1699724249) => 0.5625, //levi breath
@@ -40,12 +40,12 @@ pub fn year_2_perks() {
                 burst_delay_scale: delay,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::ExplosiveHead,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.pvp {
                 DamageModifierResponse::default()
             } else {
@@ -55,25 +55,23 @@ pub fn year_2_perks() {
                     crit_scale: 1.0,
                 }
             }
-        }),
+        },
     );
 
     add_epr(
         Perks::ExplosiveHead,
-        Box::new(
-            |_input: ModifierResponseInput| -> ExplosivePercentResponse {
-                ExplosivePercentResponse {
-                    percent: 0.5,
-                    delyed: if _input.pvp { 0.0 } else { 0.2 },
-                    retain_base_total: true,
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> ExplosivePercentResponse {
+            ExplosivePercentResponse {
+                percent: 0.5,
+                delyed: if _input.pvp { 0.0 } else { 0.2 },
+                retain_base_total: true,
+            }
+        },
     );
 
     add_rsmr(
         Perks::FeedingFrenzy,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let val = clamp(_input.value, 0, 5);
             let duration = 3.5;
             let mut reload_mult = 1.0;
@@ -102,12 +100,12 @@ pub fn year_2_perks() {
                 reload_stat_add: reload,
                 reload_time_scale: reload_mult,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::FeedingFrenzy,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut stats = HashMap::new();
             let val = clamp(_input.value, 0, 5);
             let duration = 3.5;
@@ -128,12 +126,12 @@ pub fn year_2_perks() {
             };
             stats.insert(StatHashes::RELOAD.into(), reload);
             stats
-        }),
+        },
     );
 
     add_dmr(
         Perks::FiringLine,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut crit_mult = 1.0;
             if _input.value > 0 {
                 crit_mult = 1.2;
@@ -143,24 +141,24 @@ pub fn year_2_perks() {
                 explosive_dmg_scale: 1.0,
                 impact_dmg_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_rr(
         Perks::FourthTimesTheCharm,
-        Box::new(|_input: ModifierResponseInput| -> RefundResponse {
+        |_input: ModifierResponseInput| -> RefundResponse {
             RefundResponse {
                 crit: true,
                 requirement: 4,
                 refund_mag: 2,
                 refund_reserves: 0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::KillingTally,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let val = clamp(_input.value, 0, 3);
             let mut damage_mult = 0.1 * val as f64;
             if _input.pvp {
@@ -174,32 +172,30 @@ pub fn year_2_perks() {
                 explosive_dmg_scale: 1.0 + damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_mmr(
         Perks::OverFlow,
-        Box::new(
-            |_input: ModifierResponseInput| -> MagazineModifierResponse {
-                let mut mag_scale = if _input.value > 0 { 2.0 } else { 1.0 };
-                if _input.is_enhanced && _input.value > 0 {
-                    mag_scale *= 1.1;
-                };
-                if _input.calc_data.total_shots_fired > 0.0 {
-                    mag_scale = 1.0;
-                };
-                MagazineModifierResponse {
-                    magazine_stat_add: 0,
-                    magazine_scale: mag_scale,
-                    magazine_add: 0.0,
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> MagazineModifierResponse {
+            let mut mag_scale = if _input.value > 0 { 2.0 } else { 1.0 };
+            if _input.is_enhanced && _input.value > 0 {
+                mag_scale *= 1.1;
+            };
+            if _input.calc_data.total_shots_fired > 0.0 {
+                mag_scale = 1.0;
+            };
+            MagazineModifierResponse {
+                magazine_stat_add: 0,
+                magazine_scale: mag_scale,
+                magazine_add: 0.0,
+            }
+        },
     );
 
     add_rsmr(
         Perks::RapidHit,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let values = [
                 (0, 1.0),
                 (5, 0.99),
@@ -217,12 +213,12 @@ pub fn year_2_perks() {
                 reload_stat_add: values[entry_to_get as usize].0,
                 reload_time_scale: values[entry_to_get as usize].1,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::RapidHit,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let rel_values = [0, 5, 30, 35, 45, 60];
             let stab_values = [0, 2, 12, 14, 18, 25];
             let entry_to_get = clamp(
@@ -237,12 +233,12 @@ pub fn year_2_perks() {
                 stab_values[entry_to_get as usize],
             );
             stats
-        }),
+        },
     );
 
     add_dmr(
         Perks::ResevoirBurst,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let damage_mult = if _input.calc_data.curr_mag >= _input.calc_data.base_mag {
                 1.25
             } else {
@@ -253,12 +249,12 @@ pub fn year_2_perks() {
                 explosive_dmg_scale: damage_mult,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::Surrounded,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut damage_mult = 1.0;
             if _input.value > 0 {
                 damage_mult = if *_input.calc_data.weapon_type == WeaponType::SWORD {
@@ -275,12 +271,12 @@ pub fn year_2_perks() {
                 explosive_dmg_scale: damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::FullCourt,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut damage_mult = 1.0;
             if _input.value > 0 {
                 damage_mult = 1.25;
@@ -290,12 +286,12 @@ pub fn year_2_perks() {
                 explosive_dmg_scale: damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::Swashbuckler,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let val = clamp(_input.value, 0, 5);
             let duration = if _input.is_enhanced { 6.0 } else { 4.5 };
             let mut dmg_boost = 0.067 * val as f64;
@@ -307,12 +303,12 @@ pub fn year_2_perks() {
                 explosive_dmg_scale: 1.0 + dmg_boost,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::MultikillClip,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let val = clamp(_input.value, 0, 5);
             let mut damage_mult = (1.0 / 6.0) * val as f64;
             if _input.calc_data.num_reloads > 0.0 {
@@ -323,12 +319,12 @@ pub fn year_2_perks() {
                 explosive_dmg_scale: 1.0 + damage_mult,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::ExplosiveLight,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let shots = if _input.is_enhanced { 7.0 } else { 6.0 };
             let shots_left = _input.value as f64 * shots - _input.calc_data.total_shots_fired;
             if shots_left <= 0.0 {
@@ -345,55 +341,53 @@ pub fn year_2_perks() {
                 impact_dmg_scale: mult.1,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::ExplosiveLight,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut out = HashMap::new();
             if _input.value > 0 {
                 out.insert(StatHashes::BLAST_RADIUS.into(), 100);
             };
             out
-        }),
+        },
     );
 
     add_sbr(
         Perks::EyeOfTheStorm,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut out = HashMap::new();
             if _input.value > 0 {
                 out.insert(StatHashes::HANDLING.into(), 30);
             };
             out
-        }),
+        },
     );
 
     add_hmr(
         Perks::EyeOfTheStorm,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.value > 0 {
-                    HandlingModifierResponse {
-                        stat_add: 30,
-                        ..Default::default()
-                    }
-                } else {
-                    HandlingModifierResponse::default()
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.value > 0 {
+                HandlingModifierResponse {
+                    stat_add: 30,
+                    ..Default::default()
                 }
-            },
-        ),
+            } else {
+                HandlingModifierResponse::default()
+            }
+        },
     );
 
     add_flmr(
         Perks::NoDistractions,
-        Box::new(|_input: ModifierResponseInput| -> FlinchModifierResponse {
+        |_input: ModifierResponseInput| -> FlinchModifierResponse {
             if _input.value > 0 {
                 FlinchModifierResponse { flinch_scale: 0.65 }
             } else {
                 FlinchModifierResponse::default()
             }
-        }),
+        },
     );
 }
