+++ src/perks/year_5_perks.rs
@@ -15,7 +15,7 @@ use super::{
 pub fn year_5_perks() {
     add_fmr(
         Perks::CascadePoint,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             let duration = if _input.is_enhanced { 3.0 } else { 2.5 };
             let mut delay_mult = 1.0;
             if _input.calc_data.time_total < duration && _input.value > 0 {
@@ -33,12 +33,12 @@ pub fn year_5_perks() {
                 inner_burst_scale: 1.0,
                 burst_size_add: 0.0,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::Encore,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             let val = clamp(_input.value, 0, 4) as i32;
             let stability_boost = 8 * val;
@@ -46,25 +46,24 @@ pub fn year_5_perks() {
             map.insert(StatHashes::RANGE.into(), range_boost);
             map.insert(StatHashes::STABILITY.into(), stability_boost);
             map
-        }),
+        },
     );
 
     add_rmr(
         Perks::Encore,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let val = clamp(_input.value, 0, 4) as i32;
             let range_boost = 5 * val;
             RangeModifierResponse {
                 range_stat_add: range_boost,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::FocusedFury,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
-            
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let shots_needed = (_input.calc_data.base_mag
                 * (_input.calc_data.curr_firing_data.burst_size as f64))
                 / 2.0;
@@ -80,12 +79,12 @@ pub fn year_5_perks() {
                 explosive_dmg_scale: dmg_boost,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_rmr(
         Perks::FragileFocus,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let range_bonus = if _input.value > 0 { 20 } else { 0 };
             RangeModifierResponse {
                 range_stat_add: range_bonus,
@@ -93,12 +92,12 @@ pub fn year_5_perks() {
                 range_hip_scale: 1.0,
                 range_zoom_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::FragileFocus,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             let mut range_bonus = 0;
             if _input.value > 0 {
@@ -106,12 +105,12 @@ pub fn year_5_perks() {
             };
             map.insert(StatHashes::RANGE.into(), range_bonus);
             map
-        }),
+        },
     );
 
     add_dmr(
         Perks::GutShot,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let high_weapons = [
                 WeaponType::AUTORIFLE,
                 WeaponType::HANDCANNON,
@@ -135,12 +134,12 @@ pub fn year_5_perks() {
                 explosive_dmg_scale: dmg_scale,
                 crit_scale,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::OffhandStrike,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             let mut stability_boost = 0;
             if _input.value > 0 {
@@ -148,12 +147,12 @@ pub fn year_5_perks() {
             };
             map.insert(StatHashes::STABILITY.into(), stability_boost);
             map
-        }),
+        },
     );
 
     add_rmr(
         Perks::OffhandStrike,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let mut range_hip_mult = 1.0;
             if _input.value > 0 {
                 range_hip_mult = 1.45;
@@ -164,35 +163,33 @@ pub fn year_5_perks() {
                 range_hip_scale: range_hip_mult,
                 range_zoom_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_hmr(
         Perks::Slickdraw,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                HandlingModifierResponse {
-                    stat_add: 100,
-                    stow_scale: 1.0,
-                    draw_scale: 0.95,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            HandlingModifierResponse {
+                stat_add: 100,
+                stow_scale: 1.0,
+                draw_scale: 0.95,
+                ..Default::default()
+            }
+        },
     );
 
     add_sbr(
         Perks::Slickdraw,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             map.insert(StatHashes::HANDLING.into(), 100);
             map
-        }),
+        },
     );
 
     add_sbr(
         Perks::StatsForAll,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut out = HashMap::new();
             let mut stability_boost = 0;
             let mut range_boost = 0;
@@ -209,29 +206,27 @@ pub fn year_5_perks() {
             out.insert(StatHashes::RELOAD.into(), reload_boost);
             out.insert(StatHashes::HANDLING.into(), handling_boost);
             out
-        }),
+        },
     );
 
     add_hmr(
         Perks::StatsForAll,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let mut handling_boost = 0;
-                let duration = if _input.is_enhanced { 11.0 } else { 10.0 };
-                if _input.value > 0 && _input.calc_data.time_total < duration {
-                    handling_boost = 35;
-                };
-                HandlingModifierResponse {
-                    stat_add: handling_boost,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let mut handling_boost = 0;
+            let duration = if _input.is_enhanced { 11.0 } else { 10.0 };
+            if _input.value > 0 && _input.calc_data.time_total < duration {
+                handling_boost = 35;
+            };
+            HandlingModifierResponse {
+                stat_add: handling_boost,
+                ..Default::default()
+            }
+        },
     );
 
     add_rmr(
         Perks::StatsForAll,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let mut range = 0;
             let mut range_mult = 1.0;
             if _input.value > 0 {
@@ -244,12 +239,12 @@ pub fn year_5_perks() {
                 range_hip_scale: 1.0,
                 range_zoom_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_rsmr(
         Perks::StatsForAll,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let mut reload = 0;
             let mut reload_mult = 1.0;
             let duration = if _input.is_enhanced { 11.0 } else { 10.0 };
@@ -261,12 +256,12 @@ pub fn year_5_perks() {
                 reload_stat_add: reload,
                 reload_time_scale: reload_mult,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::SteadyHands,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             let mut handling = 0;
             if _input.value > 0 {
@@ -274,33 +269,31 @@ pub fn year_5_perks() {
             };
             map.insert(StatHashes::HANDLING.into(), handling);
             map
-        }),
+        },
     );
 
     add_hmr(
         Perks::SteadyHands,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let mut handling_mult = 1.0;
-                let mut handling = 0;
-                let duration = if _input.is_enhanced { 9.0 } else { 8.5 };
-                if _input.value > 0 && _input.calc_data.time_total < duration {
-                    handling_mult = 0.825;
-                    handling = 100;
-                };
-                HandlingModifierResponse {
-                    stat_add: handling,
-                    stow_scale: handling_mult,
-                    draw_scale: handling_mult,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let mut handling_mult = 1.0;
+            let mut handling = 0;
+            let duration = if _input.is_enhanced { 9.0 } else { 8.5 };
+            if _input.value > 0 && _input.calc_data.time_total < duration {
+                handling_mult = 0.825;
+                handling = 100;
+            };
+            HandlingModifierResponse {
+                stat_add: handling,
+                stow_scale: handling_mult,
+                draw_scale: handling_mult,
+                ..Default::default()
+            }
+        },
     );
 
     add_dmr(
         Perks::TargetLock,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let nerf = 0.625; //patch 7.1.5
             let enh_increase = if _input.is_enhanced { 1.2 } else { 1.0 };
             let low_end_dmg = 0.0934 * enh_increase * nerf;
@@ -326,12 +319,12 @@ pub fn year_5_perks() {
                 explosive_dmg_scale: buff + 1.0,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::OverUnder,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let mut buff = 1.0_f64;
             if _input.calc_data.has_overshield {
                 buff += 0.2;
@@ -344,12 +337,12 @@ pub fn year_5_perks() {
                 explosive_dmg_scale: buff,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::WellRounded,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let val = clamp(_input.value, 0, 2) as i32;
             let mut map = HashMap::new();
             let stat_base = if _input.is_enhanced { 12 } else { 10 };
@@ -358,29 +351,27 @@ pub fn year_5_perks() {
             map.insert(StatHashes::RANGE.into(), stat_bump);
             map.insert(StatHashes::HANDLING.into(), stat_bump);
             map
-        }),
+        },
     );
 
     add_hmr(
         Perks::WellRounded,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let val = clamp(_input.value, 0, 2) as i32;
-                //due to ease of activation and upkeep will assume its always active
-                // let mut duration = if  _input.is_enhanced {9.0} else {8.5};
-                let stat_base = if _input.is_enhanced { 12 } else { 10 };
-                let handling = stat_base * val;
-                HandlingModifierResponse {
-                    stat_add: handling,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let val = clamp(_input.value, 0, 2) as i32;
+            //due to ease of activation and upkeep will assume its always active
+            // let mut duration = if  _input.is_enhanced {9.0} else {8.5};
+            let stat_base = if _input.is_enhanced { 12 } else { 10 };
+            let handling = stat_base * val;
+            HandlingModifierResponse {
+                stat_add: handling,
+                ..Default::default()
+            }
+        },
     );
 
     add_rmr(
         Perks::WellRounded,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let val = clamp(_input.value, 0, 2) as i32;
             let stat_base = if _input.is_enhanced { 12 } else { 10 };
             let range = stat_base * val;
@@ -390,12 +381,12 @@ pub fn year_5_perks() {
                 range_hip_scale: 1.0,
                 range_zoom_scale: 1.0,
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::BaitAndSwitch,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value > 0 {
                 DamageModifierResponse {
                     impact_dmg_scale: 1.35,
@@ -405,12 +396,12 @@ pub fn year_5_perks() {
             } else {
                 DamageModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_rsmr(
         Perks::CompulsiveReloader,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let reload_add = if _input.is_enhanced { 55 } else { 50 };
             if _input.calc_data.shots_fired_this_mag <= _input.calc_data.base_mag / 2.0
                 && _input.value > 0
@@ -422,12 +413,12 @@ pub fn year_5_perks() {
             } else {
                 ReloadModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::CompulsiveReloader,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let reload_add = if _input.is_enhanced { 55 } else { 50 };
             let mut map = HashMap::new();
             if _input.calc_data.shots_fired_this_mag <= _input.calc_data.base_mag / 2.0
@@ -436,12 +427,12 @@ pub fn year_5_perks() {
                 map.insert(StatHashes::RELOAD.into(), reload_add);
             }
             map
-        }),
+        },
     );
 
     add_sbr(
         Perks::SleightOfHand,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let val = clamp(_input.value, 0, 3) as i32;
             let mut map = HashMap::new();
             let stat_base = 10;
@@ -450,27 +441,25 @@ pub fn year_5_perks() {
             map.insert(StatHashes::RELOAD.into(), stat_bump);
             map.insert(StatHashes::HANDLING.into(), stat_bump);
             map
-        }),
+        },
     );
 
     add_hmr(
         Perks::SleightOfHand,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let val = clamp(_input.value, 0, 3) as i32;
-                let stat_base = 10;
-                let handling = stat_base * val;
-                HandlingModifierResponse {
-                    stat_add: handling,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let val = clamp(_input.value, 0, 3) as i32;
+            let stat_base = 10;
+            let handling = stat_base * val;
+            HandlingModifierResponse {
+                stat_add: handling,
+                ..Default::default()
+            }
+        },
     );
 
     add_rsmr(
         Perks::SleightOfHand,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let val = clamp(_input.value, 0, 3) as i32;
             let stat_base = 10;
             let reload = stat_base * val;
@@ -478,44 +467,42 @@ pub fn year_5_perks() {
                 reload_stat_add: reload,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_hmr(
         Perks::ShotSwap,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let mut handling_mult = 1.0;
-                let mut handling = 0;
-                if _input.value > 0 {
-                    handling_mult = 0.95;
-                    handling = 100;
-                };
-                HandlingModifierResponse {
-                    draw_add: handling,
-                    stow_add: handling,
-                    stow_scale: handling_mult,
-                    draw_scale: handling_mult,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let mut handling_mult = 1.0;
+            let mut handling = 0;
+            if _input.value > 0 {
+                handling_mult = 0.95;
+                handling = 100;
+            };
+            HandlingModifierResponse {
+                draw_add: handling,
+                stow_add: handling,
+                stow_scale: handling_mult,
+                draw_scale: handling_mult,
+                ..Default::default()
+            }
+        },
     );
 
     add_sbr(
         Perks::ShotSwap,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             if _input.value > 0 {
                 map.insert(StatHashes::HANDLING.into(), 100);
             }
             map
-        }),
+        },
     );
 
     add_fmr(
         Perks::SuccesfulWarmup,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             let fire_rate_buff = if _input.value > 0 { 0.625 } else { 1.0 };
             let duration = if _input.value > 0 {
                 6_f64
@@ -532,17 +519,17 @@ pub fn year_5_perks() {
             } else {
                 FiringModifierResponse::default()
             }
-        }),
+        },
     );
     add_dmr(
         Perks::UnstoppableForce,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let scalar = if _input.value >= 1 { 1.20 } else { 1.0 };
             DamageModifierResponse {
                 impact_dmg_scale: scalar,
                 explosive_dmg_scale: scalar,
                 crit_scale: 1.0,
             }
-        }),
+        },
     );
 }
