+++ src/perks/year_6_perks.rs
@@ -17,7 +17,7 @@ use super::{
 pub fn year_6_perks() {
     add_sbr(
         Perks::KeepAway,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             let mut range_bonus = 0;
             let mut reload_bonus = 0;
@@ -29,36 +29,36 @@ pub fn year_6_perks() {
             map.insert(StatHashes::RANGE.into(), range_bonus);
             map.insert(StatHashes::RELOAD.into(), reload_bonus);
             map
-        }),
+        },
     );
 
     add_rmr(
         Perks::KeepAway,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let ev = if _input.is_enhanced { 2 } else { 0 };
             let range_bonus = if _input.value > 0 { 10 + ev } else { 0 };
             RangeModifierResponse {
                 range_stat_add: range_bonus,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_rsmr(
         Perks::KeepAway,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let ev = if _input.is_enhanced { 2 } else { 0 };
             let reload_bonus = if _input.value > 0 { 30 + ev } else { 0 };
             ReloadModifierResponse {
                 reload_stat_add: reload_bonus,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::FieldTested,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut map = HashMap::new();
             if _input.value > 4 {
                 map.insert(StatHashes::RANGE.into(), 20);
@@ -77,12 +77,12 @@ pub fn year_6_perks() {
                 map.insert(StatHashes::RANGE.into(), 3);
             }
             map
-        }),
+        },
     );
 
     // add_hmr(
     //     Perks::FieldTested,
-    //     Box::new(
+    //
     //         |_input: ModifierResponseInput| -> HandlingModifierResponse {
     //             let val = clamp(_input.value, 0, 5) as i32;
     //             HandlingModifierResponse {
@@ -95,7 +95,7 @@ pub fn year_6_perks() {
 
     add_rsmr(
         Perks::FieldTested,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let reload_bump;
             if _input.value > 4 {
                 reload_bump = 55;
@@ -114,12 +114,12 @@ pub fn year_6_perks() {
                 reload_stat_add: reload_bump,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_rmr(
         Perks::FieldTested,
-        Box::new(|_input: ModifierResponseInput| -> RangeModifierResponse {
+        |_input: ModifierResponseInput| -> RangeModifierResponse {
             let range_bump;
             if _input.value > 4 {
                 range_bump = 20;
@@ -138,12 +138,12 @@ pub fn year_6_perks() {
                 range_stat_add: range_bump,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::ParacausalAffinity,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value > 0 {
                 DamageModifierResponse {
                     explosive_dmg_scale: 1.2,
@@ -153,29 +153,27 @@ pub fn year_6_perks() {
             } else {
                 DamageModifierResponse::default()
             }
-        }),
+        },
     );
 
     add_mmr(
         Perks::EnviousAssasin,
-        Box::new(
-            |_input: ModifierResponseInput| -> MagazineModifierResponse {
-                let val = _input.value as f64;
-                //i dont know why this if is here? - harm
-                if _input.calc_data.total_shots_fired == 0.0 {
-                    return MagazineModifierResponse {
-                        magazine_scale: 1.0 + clamp(0.1 * val, 0.0, 1.5),
-                        ..Default::default()
-                    };
+        |_input: ModifierResponseInput| -> MagazineModifierResponse {
+            let val = _input.value as f64;
+            //i dont know why this if is here? - harm
+            if _input.calc_data.total_shots_fired == 0.0 {
+                return MagazineModifierResponse {
+                    magazine_scale: 1.0 + clamp(0.1 * val, 0.0, 1.5),
+                    ..Default::default()
                 };
-                MagazineModifierResponse::default()
-            },
-        ),
+            };
+            MagazineModifierResponse::default()
+        },
     );
 
     add_dmr(
         Perks::CollectiveAction,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = match (_input.pvp, _input.value) {
                 (_, 0) => 1.0,
                 (true, 1..) => 1.1,
@@ -186,57 +184,53 @@ pub fn year_6_perks() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_dmr(
         Perks::Bipod,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             DamageModifierResponse {
                 impact_dmg_scale: 0.75,
                 explosive_dmg_scale: 0.75,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_mmr(
         Perks::Bipod,
-        Box::new(
-            |_input: ModifierResponseInput| -> MagazineModifierResponse {
-                MagazineModifierResponse {
-                    magazine_scale: 2.0,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> MagazineModifierResponse {
+            MagazineModifierResponse {
+                magazine_scale: 2.0,
+                ..Default::default()
+            }
+        },
     );
 
     add_imr(
         Perks::Bipod,
-        Box::new(
-            |_input: ModifierResponseInput| -> InventoryModifierResponse {
-                InventoryModifierResponse {
-                    inv_add: 5,
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> InventoryModifierResponse {
+            InventoryModifierResponse {
+                inv_add: 5,
+                ..Default::default()
+            }
+        },
     );
 
     add_fmr(
         Perks::Bipod,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             FiringModifierResponse {
                 burst_delay_scale: 0.75,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_fmr(
         Perks::ControlledBurst,
-        Box::new(|_input: ModifierResponseInput| -> FiringModifierResponse {
+        |_input: ModifierResponseInput| -> FiringModifierResponse {
             if _input.value > 0 {
                 return FiringModifierResponse {
                     burst_delay_scale: 0.9,
@@ -244,11 +238,11 @@ pub fn year_6_perks() {
                 };
             }
             FiringModifierResponse::default()
-        }),
+        },
     );
     add_dmr(
         Perks::ControlledBurst,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value > 0 {
                 return DamageModifierResponse {
                     impact_dmg_scale: 1.2,
@@ -257,21 +251,21 @@ pub fn year_6_perks() {
                 };
             }
             DamageModifierResponse::default()
-        }),
+        },
     );
     add_sbr(
         Perks::InvisibleHand,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut stats = HashMap::new();
             if _input.value > 0 {
                 stats.insert(StatHashes::STABILITY.into(), 25);
             }
             stats
-        }),
+        },
     );
     add_sbr(
         Perks::UnsatedHunger,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut stats = HashMap::new();
             if _input.value > 0 {
                 stats.insert(StatHashes::STABILITY.into(), 20);
@@ -279,12 +273,12 @@ pub fn year_6_perks() {
                 stats.insert(StatHashes::RELOAD.into(), 60);
             }
             stats
-        }),
+        },
     );
 
     add_rsmr(
         Perks::UnsatedHunger,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             if _input.value > 0 {
                 return ReloadModifierResponse {
                     reload_stat_add: 60,
@@ -292,52 +286,48 @@ pub fn year_6_perks() {
                 };
             }
             ReloadModifierResponse::default()
-        }),
+        },
     );
     add_hmr(
         Perks::UnsatedHunger,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.value > 0 {
-                    return HandlingModifierResponse {
-                        stat_add: 60,
-                        ..Default::default()
-                    };
-                }
-                HandlingModifierResponse::default()
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.value > 0 {
+                return HandlingModifierResponse {
+                    stat_add: 60,
+                    ..Default::default()
+                };
+            }
+            HandlingModifierResponse::default()
+        },
     );
 
     add_hmr(
         Perks::Discord,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.value > 0 {
-                    return HandlingModifierResponse {
-                        ads_scale: 0.75,
-                        ..Default::default()
-                    };
-                }
-                HandlingModifierResponse::default()
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.value > 0 {
+                return HandlingModifierResponse {
+                    ads_scale: 0.75,
+                    ..Default::default()
+                };
+            }
+            HandlingModifierResponse::default()
+        },
     );
 
     add_sbr(
         Perks::Discord,
-        Box::new(|_input: ModifierResponseInput| -> HashMap<u32, i32> {
+        |_input: ModifierResponseInput| -> HashMap<u32, i32> {
             let mut stats = HashMap::new();
             if _input.value > 0 {
                 stats.insert(StatHashes::AIRBORNE.into(), 30);
             }
             stats
-        }),
+        },
     );
 
     add_dmr(
         Perks::PrecisionInstrument,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let max_percent = if _input.is_enhanced { 0.26 } else { 0.25 };
             let max_stacks = 6.0;
             let shots_hit = _input.calc_data.total_shots_hit;
@@ -348,12 +338,12 @@ pub fn year_6_perks() {
                 crit_scale: 1.0 + stacks * max_percent / max_stacks,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_rsmr(
         Perks::LooseChange,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             if _input.value == 0 {
                 return ReloadModifierResponse::default();
             }
@@ -361,24 +351,22 @@ pub fn year_6_perks() {
                 reload_stat_add: 50,
                 ..Default::default()
             }
-        }),
+        },
     );
 
     add_sbr(
         Perks::LooseChange,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                if _input.value == 0 {
-                    return HashMap::new();
-                }
-                HashMap::from([(StatHashes::RELOAD.into(), 50)])
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            if _input.value == 0 {
+                return HashMap::new();
+            }
+            HashMap::from([(StatHashes::RELOAD.into(), 50)])
+        },
     );
 
     add_dmr(
         Perks::HighGround,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             if _input.value == 0 {
                 return DamageModifierResponse::default();
             }
@@ -389,39 +377,35 @@ pub fn year_6_perks() {
                 explosive_dmg_scale: mult,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_sbr(
         Perks::HeadRush,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                if _input.value == 0 {
-                    return HashMap::new();
-                }
-                HashMap::from([
-                    (StatHashes::RELOAD.into(), 10),
-                    (StatHashes::HANDLING.into(), 0),
-                ])
-            },
-        ),
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            if _input.value == 0 {
+                return HashMap::new();
+            }
+            HashMap::from([
+                (StatHashes::RELOAD.into(), 10),
+                (StatHashes::HANDLING.into(), 0),
+            ])
+        },
     );
     add_hmr(
         Perks::HeadRush,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                if _input.value == 0 {
-                    return HandlingModifierResponse::default();
-                }
-                //unknown at time
-                HandlingModifierResponse {
-                    ..Default::default()
-                }
-            },
-        ),
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            if _input.value == 0 {
+                return HandlingModifierResponse::default();
+            }
+            //unknown at time
+            HandlingModifierResponse {
+                ..Default::default()
+            }
+        },
     );
     add_rsmr(
         Perks::HeadRush,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             if _input.value == 0 {
                 return ReloadModifierResponse::default();
             }
@@ -429,45 +413,41 @@ pub fn year_6_perks() {
                 reload_stat_add: 10,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_sbr(
         Perks::EnlightendAction,
-        Box::new(
-            |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
-                let shots_hit = _input.calc_data.total_shots_hit as i32;
-                let value = _input.value as i32;
-                let stat_per_stack = 10;
-                let max_stacks = 5;
+        |_input: ModifierResponseInput| -> HashMap<BungieHash, StatBump> {
+            let shots_hit = _input.calc_data.total_shots_hit as i32;
+            let value = _input.value as i32;
+            let stat_per_stack = 10;
+            let max_stacks = 5;
 
-                let stat_bump = clamp(value + shots_hit, 0, max_stacks) * stat_per_stack;
-                HashMap::from([
-                    (StatHashes::RELOAD.into(), stat_bump),
-                    (StatHashes::HANDLING.into(), stat_bump),
-                ])
-            },
-        ),
+            let stat_bump = clamp(value + shots_hit, 0, max_stacks) * stat_per_stack;
+            HashMap::from([
+                (StatHashes::RELOAD.into(), stat_bump),
+                (StatHashes::HANDLING.into(), stat_bump),
+            ])
+        },
     );
     add_hmr(
         Perks::EnlightendAction,
-        Box::new(
-            |_input: ModifierResponseInput| -> HandlingModifierResponse {
-                let shots_hit = _input.calc_data.total_shots_hit as i32;
-                let value = _input.value as i32;
-                let stat_per_stack = 10;
-                let max_stacks = 5;
+        |_input: ModifierResponseInput| -> HandlingModifierResponse {
+            let shots_hit = _input.calc_data.total_shots_hit as i32;
+            let value = _input.value as i32;
+            let stat_per_stack = 10;
+            let max_stacks = 5;
 
-                let stat_bump = clamp(value + shots_hit, 0, max_stacks) * stat_per_stack;
-                HandlingModifierResponse {
-                    stat_add: stat_bump,
-                    ..Default::default()
-                }
-            },
-        ),
+            let stat_bump = clamp(value + shots_hit, 0, max_stacks) * stat_per_stack;
+            HandlingModifierResponse {
+                stat_add: stat_bump,
+                ..Default::default()
+            }
+        },
     );
     add_rsmr(
         Perks::EnlightendAction,
-        Box::new(|_input: ModifierResponseInput| -> ReloadModifierResponse {
+        |_input: ModifierResponseInput| -> ReloadModifierResponse {
             let shots_hit = _input.calc_data.total_shots_hit as i32;
             let value = _input.value as i32;
             let stat_per_stack = 10;
@@ -478,11 +458,11 @@ pub fn year_6_perks() {
                 reload_stat_add: stat_bump,
                 ..Default::default()
             }
-        }),
+        },
     );
     add_dmr(
         Perks::SwordLogic,
-        Box::new(|_input: ModifierResponseInput| -> DamageModifierResponse {
+        |_input: ModifierResponseInput| -> DamageModifierResponse {
             let buff = match _input.value {
                 0 => 1.0,
                 1..=3 => 1.05 + (0.1 * _input.value as f64),
@@ -493,6 +473,6 @@ pub fn year_6_perks() {
                 explosive_dmg_scale: buff,
                 ..Default::default()
             }
-        }),
+        },
     )
 }
