+++ src/types/mod.rs
@@ -1,3 +1,10 @@
 pub mod js_types;
 
 pub mod rs_types;
+pub mod formula_types;
+
+pub mod prelude {
+    pub use super::formula_types::*;
+    pub use super::js_types::*;
+    pub use super::rs_types::*;
+}
